script: Move individual files to respective subfolders
All of these structs are referenced only in a specific
subdirectory, or are closely related to them. Therefore,
to clean up the `dom/` directory, move them closer to
the other code.

Part of #38901

Testing: it compiles
