+++ components/script/dom/document/mod.rs
@@ -11,5 +11,7 @@ pub(crate) mod documentfragment;
 pub(crate) mod documentorshadowroot;
 pub(crate) mod documenttype;
 pub(crate) mod focus;
+mod tree_ordered_index_map;
+pub(crate) mod visibilitystateentry;
 
 pub(crate) use self::document::*;
