+++ components/script/dom/document/visibilitystateentry.rs
