+++ components/script/dom/element/attributes/mod.rs
@@ -3,4 +3,5 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 pub(crate) mod accessors;
+pub(crate) mod attr;
 pub(crate) mod storage;
